dont delete files if debug is set
git-svn-id: file:///var/svn/bend@122 f4574214-f815-0410-b9ce-81fe6a62d61e

+++ dvd2mkv.php
@@ -252,7 +252,7 @@
 			$dvd->mkvmerge($avi, $txt, $mkv, $title, $atrack, $aspect, $idx);
 		}
 
-		if(file_exists($mkv)) {
+		if(file_exists($mkv) && !$dvd->args['debug']) {
 			file_exists($vob) && unlink($vob);
 			file_exists($avi) && unlink($avi);
 			file_exists($txt) && unlink($txt);
